refactor(WinningNumber): 계층간 분리를 위해 WinningNumber 객체를 LottoController 에서 생성하도록 수정
